[fix] Allow user to go back to editing pad after a dialog is closed
When a dialog is being closed, we're bluring the focus on padOuter,
which does not allow user to go back to typing after dialog finishes
closing. The code that is causing the issue is mobile-specific, so we're
removing it -- a broader strategy needs to be planned anyway when we
decide to support mobile devices.