[fix] Avoid marking ALL TEXT with preText attribs (woot?!)
There was a bug on `collectContentPre`: it was not checking for
pre-existing attribs, it was simply adding the preText attrib EVERY
TIME!

As a consequence, after a very long paste, the entire pad would be
marked with the attrib, and when pad was reloaded it took forever to
clean up all those lines, which was causing the browser to freeze and
the "saving" message to be displayed several times to the user.

Fix https://trello.com/c/KskaudQ5/1190.