[fix] Only calculate icon position when target element is on pad
Avoid errors when updating the position of text mark icons: if the
target element is not on pad, don't try to get its position.

This happens, for example, when pasting a comment: the comment is pasted
with id "fake-c-...", but the `textMarkId` is "c-...", which is not yet
on the pad.

Fix error (1) of https://trello.com/c/3aUCymV2/1235.