[refactor] Extract component to trigger event every time a line changes
Create a module (`lineChangeEventTriggerer.js`) to calculate the
top-most line number that was changed on each batch of pad changes, and
trigger a unique event with that number for other plugins to use.

This avoid having the same code to calculate that line number,
duplicated around several plugins. Besides removing some lines of code,
we are also more efficient with these changes.

This is part of the fixes for bug https://trello.com/c/SgiGE4V0/1377.