[fix] Fix bug of show/hide replies button text
In this commit we fix the error that was displaying always the button
text as "hide replies". This bug was caused because the introduction of
animation. The animation has created a step between show and hide. So, for
example, in the scenario of changing from visible to hidden
the window animation would take some seconds to change its visibility.
As we query this visibility to decide if the button should have "show
replies or hide replies" as its text, we were not getting the final state
of the window