[feat] Send comment ids on "new data" API event
We need to send the ids of current comments on text because the plugin
API only provides the full set of comment data, not only the set of
comments that are actually still on text. This means that if some text
with a comment is removed from the pad, the plugin API will still return
that comment.

To allow merging the comments data with the list of comments still on
text, we provide the comment ids on the 'comments_set_changed' message
and let consumers merge that with the data from the plugin API on `GET
/p/:pad/:rev?/comments`.