[fix] Only send API event when pasted comments are fully processed (#64)
When multiple comments are pasted, it might take a while for the
comments to be available both on database and on text. As we need both
sides to be available when sending the `comments_set_changed` API event,
we need to make sure we wait for everything to be ready when we send the
data on the API, otherwise we might send the event with an empty list of
comments, as if no comment was pasted at all.

Fix https://trello.com/c/ARur9FcO/1905.