[feat] Add the afterUpdateIconsPosition function (#69)
* [feat] Add the afterUpdateIconsPosition function

This commit is related to card 2192, which implements the icons
grouping. All that grouping logic was implemented in ep_script_touches
plugin. The changes added by this commit allow to run a function, if provided,
after update the icons position. This way we can group the icons after
their positioning is finished.